fix warning

+++ test/src/extractor/test_avc_to_annex_b.cpp
@@ -34,7 +34,10 @@ TEST(extractor_test_avc_to_annex_b, init)
         test_sample.data(), test_sample.size(), nalu_size_length);
 
     EXPECT_EQ(5U, samples.size());
-    uint64_t sum_size = std::accumulate(samples.begin(), samples.end(), 0U,
+    uint64_t sum_size = (uint64_t)std::accumulate(
+        samples.begin(),
+        samples.end(),
+        0U,
         [](uint64_t accumlate, auto s){ return accumlate + s.second; });
     (void) sum_size;
     // uint64_t expected_sum_size = sample_size - (samples.size() * nalu_size_length);
